Refactor : Search V1 asynchronization
