+++ src/main/java/com/example/lucky7/config/AsyncConfig.java
@@ -1,0 +1,9 @@
+package com.example.lucky7.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.annotation.EnableAsync;
+
+@Configuration
+@EnableAsync
+public class AsyncConfig {
+}
