+++ src/main/java/com/example/lucky7/domain/search/dto/response/SearchResponse.java
@@ -4,13 +4,15 @@ import com.example.lucky7.domain.store.entity.Store;
 import com.example.lucky7.domain.store.enums.StoreCategory;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Getter
 @AllArgsConstructor
+@NoArgsConstructor
 public class SearchResponse {
-    private final Long id;
-    private final String name;
-    private final StoreCategory category;
+    private Long id;
+    private String name;
+    private StoreCategory category;
 
     public static SearchResponse toDto(Store store){
         return new SearchResponse(
