+++ src/main/java/com/example/lucky7/domain/store/controller/StoreController.java
@@ -76,22 +76,22 @@ public class StoreController {
         return ResponseEntity.noContent().build();
     }
 
-    // ------------------- GeoHash 사용한 위치 기반 검색 시작 ----------------------------
-
-    @PostMapping("/geohash")
-    public ResponseEntity<StoreResponse> saveWithGeoHash(@RequestBody StoreCreateRequest request) {
-        return ResponseEntity.ok(storeService.saveWithGeoHash(request));
-    }
-
-    @GetMapping("/geohash")
-    public ResponseEntity<List<StoreResponse>> getNearByGeoHash(
-            @RequestParam(value = "longitude") double longitude,
-            @RequestParam(value = "latitude") double latitude,
-            @RequestParam(value = "distance") double distance) {
-        List<StoreResponse> nearbyStores = storeService.findNearbyGeoHash(longitude, latitude, distance);
-        return ResponseEntity.ok(nearbyStores);
-    }
-
-    // ------------------- GeoHash 사용한 위치 기반 검색 끝 ----------------------------
+//    // ------------------- GeoHash 사용한 위치 기반 검색 시작 ----------------------------
+//
+//    @PostMapping("/geohash")
+//    public ResponseEntity<StoreResponse> saveWithGeoHash(@RequestBody StoreCreateRequest request) {
+//        return ResponseEntity.ok(storeService.saveWithGeoHash(request));
+//    }
+//
+//    @GetMapping("/geohash")
+//    public ResponseEntity<List<StoreResponse>> getNearByGeoHash(
+//            @RequestParam(value = "longitude") double longitude,
+//            @RequestParam(value = "latitude") double latitude,
+//            @RequestParam(value = "distance") double distance) {
+//        List<StoreResponse> nearbyStores = storeService.findNearbyGeoHash(longitude, latitude, distance);
+//        return ResponseEntity.ok(nearbyStores);
+//    }
+//
+//    // ------------------- GeoHash 사용한 위치 기반 검색 끝 ----------------------------
 
 }
